+++ plugins/virtual_event/lib/api/types/KalturaVirtualEventListResponse.php
@@ -3,7 +3,7 @@
  * @package plugins.virtualEvent
  * @subpackage api.types
  */
-class KalturaVirtualEventFilterListResponse extends KalturaListResponse
+class KalturaVirtualEventListResponse extends KalturaListResponse
 {
 	/**
 	 * @var KalturaVirtualEventArray